stub API with JSON files
