+++ apps/mattermost.yaml
@@ -1,0 +1,9 @@
+kind: Application
+metadata:
+  name: mattermost
+spec:
+  repository: ghcr.io/defenseunicorns/packages/uds/mattermost
+  links:
+    - description: UDS Package
+      url: "https://github.com/defenseunicorns/uds-package-mattermost"
+  versions: []
