+++ apps/sonarqube.yaml
@@ -1,0 +1,9 @@
+kind: Application
+metadata:
+  name: sonarqube
+spec:
+  repository: ghcr.io/defenseunicorns/packages/uds/sonarqube
+  links:
+    - description: UDS Package
+      url: "https://github.com/defenseunicorns/uds-package-sonarqube"
+  versions: []
