+++ go.mod
@@ -8,6 +8,14 @@ require (
 )
 
 require (
+	github.com/alecthomas/jsonschema v0.0.0-20220216202328-9eeeec9d044b // indirect
+	github.com/defenseunicorns/uds-cli v0.14.0 // indirect
+	github.com/iancoleman/orderedmap v0.0.0-20190318233801-ac98e3ecb4b0 // indirect
+	k8s.io/apimachinery v0.30.3 // indirect
+	k8s.io/utils v0.0.0-20231127182322-b307cd553661 // indirect
+)
+
+require (
 	atomicgo.dev/cursor v0.2.0 // indirect
 	atomicgo.dev/keyboard v0.2.9 // indirect
 	atomicgo.dev/schedule v0.1.0 // indirect
@@ -15,7 +23,7 @@ require (
 	github.com/buger/jsonparser v1.1.1 // indirect
 	github.com/containerd/console v1.0.4-0.20230313162750-1ae8d489ac81 // indirect
 	github.com/defenseunicorns/pkg/helpers/v2 v2.0.1 // indirect
-	github.com/fatih/color v1.16.0 // indirect
+	github.com/fatih/color v1.17.0 // indirect
 	github.com/gookit/color v1.5.4 // indirect
 	github.com/invopop/jsonschema v0.12.0 // indirect
 	github.com/kr/pretty v0.3.1 // indirect
@@ -25,11 +33,12 @@ require (
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mattn/go-runewidth v0.0.15 // indirect
 	github.com/otiai10/copy v1.14.0 // indirect
-	github.com/pterm/pterm v0.12.78 // indirect
-	github.com/rivo/uniseg v0.4.4 // indirect
+	github.com/pterm/pterm v0.12.79 // indirect
+	github.com/rivo/uniseg v0.4.7 // indirect
 	github.com/rogpeppe/go-internal v1.12.0 // indirect
 	github.com/wk8/go-ordered-map/v2 v2.1.8 // indirect
 	github.com/xo/terminfo v0.0.0-20220910002029-abceb7e1c41e // indirect
+	github.com/zarf-dev/zarf/src/api v0.0.0-20240808165401-9f442cea1f98
 	golang.org/x/sync v0.7.0 // indirect
 	golang.org/x/sys v0.22.0 // indirect
 	golang.org/x/term v0.22.0 // indirect
