--- ui/src/data/packages/uds/gitlab/zarf.json
@@ -1,127 +1,0 @@
-{
-  "kind": "ZarfPackageConfig",
-  "metadata": {
-    "name": "gitlab",
-    "description": "UDS GitLab Package",
-    "version": "17.2.1-uds.0",
-    "architecture": "amd64",
-    "aggregateChecksum": "458f6879804ba63197c9ae559a5e9869acc5a71eb6b39d7bc77f31dc675e3599"
-  },
-  "build": {
-    "terminal": "runner",
-    "user": "runner",
-    "architecture": "amd64",
-    "timestamp": "Mon, 29 Jul 2024 22:45:07 +0000",
-    "version": "v0.36.1",
-    "migrations": [
-      "scripts-to-actions",
-      "pluralize-set-variable"
-    ],
-    "lastNonBreakingVersion": "v0.27.0",
-    "flavor": "registry1"
-  },
-  "components": [
-    {
-      "name": "gitlab",
-      "description": "Deploy gitlab with registry1 images",
-      "required": true,
-      "charts": [
-        {
-          "name": "uds-gitlab-config",
-          "version": "0.2.0",
-          "localPath": "chart",
-          "namespace": "gitlab"
-        },
-        {
-          "name": "gitlab",
-          "version": "8.1.3",
-          "url": "https://charts.gitlab.io/",
-          "gitPath": "chart",
-          "namespace": "gitlab",
-          "valuesFiles": [
-            "values/common-values.yaml",
-            "values/registry1-values.yaml"
-          ]
-        }
-      ],
-      "images": [
-        "registry1.dso.mil/ironbank/gitlab/gitlab/certificates:17.2.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/cfssl-self-sign:1.6.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/gitaly:17.2.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-container-registry:17.2.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-pages:17.2.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-shell:17.2.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-sidekiq:17.2.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-toolbox:17.2.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-webservice:17.2.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-workhorse:17.2.1",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/kubectl:17.2.1",
-        "registry1.dso.mil/ironbank/redhat/ubi/ubi9:9.4",
-        "registry1.dso.mil/ironbank/gitlab/gitlab/gitlab-exporter:17.2.1"
-      ],
-      "actions": {
-        "onDeploy": {
-          "after": [
-            {
-              "maxTotalSeconds": 300,
-              "description": "Validate GitLab Package",
-              "wait": {
-                "cluster": {
-                  "kind": "Packages",
-                  "name": "gitlab",
-                  "namespace": "gitlab",
-                  "condition": "'{.status.phase}'=Ready"
-                }
-              }
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "variables": [
-    {
-      "name": "DOMAIN",
-      "default": "uds.dev"
-    },
-    {
-      "name": "BUCKET_PREFIX",
-      "default": "uds-"
-    },
-    {
-      "name": "BUCKET_SUFFIX"
-    },
-    {
-      "name": "GITLAB_REDIS_ENDPOINT"
-    },
-    {
-      "name": "GITLAB_REDIS_SCHEME",
-      "default": "redis"
-    },
-    {
-      "name": "GITLAB_DB_NAME",
-      "default": "gitlabdb"
-    },
-    {
-      "name": "GITLAB_DB_USERNAME",
-      "default": "gitlab"
-    },
-    {
-      "name": "GITLAB_DB_ENDPOINT",
-      "default": "postgresql"
-    },
-    {
-      "name": "GITLAB_PAGES_ENABLED",
-      "default": "false"
-    },
-    {
-      "name": "GITLAB_SIGNUP_ENABLED",
-      "default": "true"
-    },
-    {
-      "name": "DISABLE_REGISTRY_REDIRECT",
-      "description": "If your storage endpoint is not publicly accessible set this to true",
-      "default": "false"
-    }
-  ]
-}
