+++ ui/src/types/index.ts
@@ -1,0 +1,18 @@
+export * from './gen';
+
+export interface Application {
+  kind: 'Application';
+  meta: {
+    name: string;
+  };
+  spec: {
+    description: string;
+    repository: string;
+    links: {
+      description: string;
+      url: string;
+    }[];
+    keywords: string[];
+    versions: string[];
+  }
+}
